@@ -14,12 +14,12 @@
         }
     ],
     "require": {
-        "php": "^7.1.0 || ^8.0",
-        "sabre/vobject": "^4.2.1",
-        "sabre/event" : "^5.0",
-        "sabre/xml"  : "^2.0.1",
-        "sabre/http" : "^5.0.5",
-        "sabre/uri" : "^2.0",
+        "php": "^7.4 || ^8.0",
+        "sabre/vobject": "^4.5.2",
+        "sabre/event" : "^6.0",
+        "sabre/xml"  : "^3.0 || ^4.0",
+        "sabre/http" : "^6.0 || ^7.0",
+        "sabre/uri" : "^2.3 || ^3.0",
         "ext-dom": "*",
         "ext-pcre": "*",
         "ext-spl": "*",
@@ -33,10 +33,10 @@
         "ext-json": "*"
     },
     "require-dev" : {
-        "friendsofphp/php-cs-fixer": "^2.17.1",
-        "phpstan/phpstan": "^0.12",
-        "phpunit/phpunit" : "^7.5 || ^8.5 || ^9.0",
-        "monolog/monolog": "^1.18"
+        "friendsofphp/php-cs-fixer": "^3.14",
+        "phpstan/phpstan": "^1.9",
+        "phpunit/phpunit" : "^9.0",
+        "monolog/monolog": "^2.8"
     },
     "suggest" : {
         "ext-curl" : "*",
