@@ -62,7 +62,7 @@ class EmailAddressSet implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->emails as $email) {
             $writer->writeElement('{http://calendarserver.org/ns/}email-address', $email);
