@@ -68,7 +68,7 @@ class Invite implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         $cs = '{'.Plugin::NS_CALENDARSERVER.'}';
 
