@@ -77,7 +77,7 @@ class ScheduleCalendarTransp implements Element
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         switch ($this->value) {
             case self::TRANSPARENT:
