@@ -68,7 +68,7 @@ class SupportedCalendarComponentSet implements Element
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->components as $component) {
             $writer->startElement('{'.Plugin::NS_CALDAV.'}comp');
