@@ -40,7 +40,7 @@ class SupportedCalendarData implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         $writer->startElement('{'.Plugin::NS_CALDAV.'}calendar-data');
         $writer->writeAttributes([
