@@ -39,7 +39,7 @@ class SupportedCollationSet implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         $collations = [
             'i;ascii-casemap',
