@@ -63,7 +63,7 @@ class SupportedAddressData implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->supportedData as $supported) {
             $writer->startElement('{'.Plugin::NS_CARDDAV.'}address-data-type');
