@@ -57,7 +57,7 @@ class Invite implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->sharees as $sharee) {
             $writer->writeElement('{DAV:}sharee', $sharee);
