@@ -63,7 +63,7 @@ class LockDiscovery implements XmlSerializable
      * Important note 2: If you are writing any new elements, you are also
      * responsible for closing them.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->locks as $lock) {
             $writer->startElement('{DAV:}activelock');
