@@ -72,7 +72,7 @@ class ShareAccess implements Element
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         switch ($this->value) {
             case SharingPlugin::ACCESS_NOTSHARED:
