@@ -82,7 +82,7 @@ class SupportedMethodSet implements XmlSerializable, HtmlOutput
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->getValue() as $val) {
             $writer->startElement('{DAV:}supported-method');
