@@ -92,7 +92,7 @@ class Acl implements Element, HtmlOutput
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         foreach ($this->privileges as $ace) {
             $this->serializeAce($writer, $ace);
