@@ -34,7 +34,7 @@ class AclRestrictions implements XmlSerializable
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         $writer->writeElement('{DAV:}grant-only');
         $writer->writeElement('{DAV:}no-invert');
