@@ -99,7 +99,7 @@ class Principal extends DAV\Xml\Property\Href
      *
      * If you are opening new elements, you must also close them again.
      */
-    public function xmlSerialize(Writer $writer)
+    public function xmlSerialize(Writer $writer): void
     {
         switch ($this->type) {
             case self::UNAUTHENTICATED:
