@@ -24,7 +24,7 @@ class MockLogger extends AbstractLogger
      * @param mixed  $level
      * @param string $message
      */
-    public function log($level, $message, array $context = [])
+    public function log($level, $message, array $context = []): void
     {
         $this->logs[] = [
             $level,
